DEV-39056 - Fix scroll wheel on prims
the scrollEvent() API's x and y are not the same as
handleScrollWheel's x and y. The latter is the position of the mouse
at the time of the event. The former is the 'scroll amount' in x and
y, respectively. All we have for 'scroll amount' here is 'clicks', and
no mask.