Fix for DEV-41978 (media controls URL is incorrect after whitelist bounces webpage back).
Added LLViewerMediaImpl::getCurrentMediaURL(), which reflects the most recent url either set with navigateTo or received from the plugin via locationChanged/navigateComplete.  LLViewerMediaImpl now uses this when reload() is called.

LLPanelPrimMediaControls now uses getCurrentMediaURL() to get the url it displays in the URL input bar and opens with the pop-out control.