EXT-1588: Remove support for automatic hyperlinking of URLs in notecards until this issue can be addressed correctly.
The best solution would be to implement support for alternative
display strings for URL text segments. That way, the raw text will
contain the original URL string, but we can show a display-only
alternate label for the URL. Until I get round to doing this, I'm
turning off URL hyperlinking in notecards to avoid the potential
for data loss.