EXT-2326 Editing appearance causes too many texture callbacks stalls the viewer
problem was caused by triggering updateMeshTextures() every time visual params
or textures were changed on your base avatar object. The latest structure
makes the parameters and textures in the wearables authoritative, causing them
to be copied to the avatar on a regular basis. This resulted in way too many callbacks
to be registered while waiting for baked textures to be uploaded.

Fixed by removing many calls to updateMeshTextures() and ensuring that it gets called
every time an appropriate param or texture actually changes, or a wearable gets
added/removed. This appears to cut the number of callbacks to a reasonable level.

Code reviewed by Bigpapi.