DEV-2925: Added a new remote object inspector.
This is used to display details about objects that may not be in the
current scene (the existing object inspector is tied to selection and
only works for objects in the scene). The remote inspector lets you see
the name and owner of the object. You can also teleport to the object's
location, view the location on the map, or mute the object.

I've also added more information to the notification you receive when
an object sends you an IM via llInstantMessage(). This notification now
has an "Inspect" button that brings up the remote object inspector.