EXT-3624: Support "_blank" href target types.
Support href links in web views that have a target attribute set to
"_blank". This will open the link in the user's preferred browser
(internal or external, as defined in Prefs). This change relies on the
new llqtwebkit package that has just been pushed.

Also updated the hardcoded Qt version number in the About floater.