EXT-4693: First steps at using remote inspector.
Now Object IMs that get sent to your local chat window will have
secondlife:///app/objectim SLapps, which will open the remote object
inspector for that object, not the avatar inspector.

However, we don't have all the information needed for the remote
object inspector, so I'm going to have to investigate extending the
inspector to request the missing information from the server.