Further fixes for EXT-4689 (Long-word chat spam cripples fps and/or disconnects client).
This should fix the inefficiencies in the append path that made viewer FPS drop severely when addinglarge amounts of text to the nearby chat floater.

Resizing the floater with a huge amount of text in it is still pretty bad, but fixing that will require some bigger architectural changes.

Changed LLTextBase::needsReflow() to take an offset at which to start reflow processing.

Changed most needsReflow() calls in LLTextBase to supply a proper index.

Changed LLTextBase::reflow() to use the reflow index maintained by needsReflow().

Removed all needsReflow() calls from LLTextEditor (the only way for it to manipulate the text is through functions in LLTextBase that already manage reflowing internally).

Removed LLTextEditor::replaceUrlLabel(), since it was identical to the inherited version LLTextBase::replaceUrlLabel().

Reviewed by Richard.