EXT-5364 New users appear fully naked before resolving.
We were pushing off redrawing your local texture composite
while waiting for a baked texture cache response from the server.
If this happens when your textures come in, your avatar will de-cloud
but the composite will not update, resulting in nudity.

Changed the logic so we will re-generate a local composite even if
we have a pending baked texture query (we still suppress uploading the
composite until a response is received). Verified several times that this does
fix the issue. Will update the JIRA with reliable repro.

Q verified issue is worthy of a hotfix checkin

Code has been reviewed by vir, bigpapi, and seraph

This is necessary for new user experience. Patch should be contained enough to be low-risk. Contact Nyx if any further questions.