Fix for EXT-6756:  google apps auth doesn't work right with shared media cookies
Added "HttpOnly" to the allowed field names in LLPluginCookieStore::Cookie::parse().  (This was the actual cause of the failure -- cookies with this field in them were silently failing to parse.)

Added some LL_WARNS logging on this sort of cookie parse failure, which will make similar problems much easier to track down in future.

Also added tags to most of the logging in llplugincookiestore.cpp to make it easier to selectively enable it when debugging.

Added a cookie with all allowable field names to the unit test.

Reviewed by Sam at http://codereview.lindenlab.com/1247014