Fix for incorrect handling of an OpenID url containing a port number
The code was using the "authority" part of the URL returned by LLURL when constructing the cookie's domain, which incorrectly included the port number.  This change makes sure any port number and username/password that might be in the authority are stripped from the string before handing it to the cookie code.

Reviewed by Gino at http://codereview.lindenlab.com/1297006