DEV-48904 - p2p acceptance dialog vanishes too quickly DEV-48903 - malformed names in p2p dialogs for Diamondware In the merge there were some changes that didn't get propagated that resulted in some failures with respect to diamondware name processing.  THe determination as to whether a p2p session was with an avatar or with an alcatel (pbx) wasn't propagated into the diamondware code.
I fixed that by merging the changes to the vivox module with respect to that into the diamondware module.

Also, the acceptance dialog was vanishing too quickly as the determination as to whether there was still
a p2p invite pending was not updated when the calling code was updated.
The calling code initially was changed to look into vivox state which isn't really valid.  that was changed in
the 'trunk' to another mechanism, but that still assumed visibility into the voice code.

I fixed that by creating an API call to ask whether a call was pending.

CR: Karina