EXT-7005 : Kill unhelpful 'Deprecated conversion from string to char *' warnings from fmod_errors.h when building with GCC >=4.2 on Darwin
Apply -Wno-write-strings compile flag to llaudioengine_fmod.cpp and llstreamingaudio_fmod.cpp on Darwin as it was on Linux (Amen).
Removed unnecessary check for GCC >= 4.2 when applying this flag, as CXX_VERSION_NUMBER it used is only being generated in 00-Common.cmake on Linux at the moment and 4.0 understands -Wno-write-strings anyway.