Fix for EXT-5667 (support window.open() in shared media so google docs will work).
Moved the processing of clicks on _external and _blank targeted links from LLMediaCtrl to LLViewerMediaImpl.

Removed LLMediaCtrl::setOpenInExternalBrowser() since that functionality is available through the use of the _external target attribute in web content.

Removed LLMediaCtrl:: setOpenInInternalBrowser() since it was unimplemented and not used.

Made the webkit media plugin set llqtwebkit's window open behavior to WOB_SIMULATE_BLANK_HREF_CLICK.  This is #ifdefed out on Linux until we get a new Linux build of llqtwebkit.