STORM-227 FIXED Made long cache location in Preferences->Setup readable.
Changes:
- The line editor is now scrollable.
- Its value is duplicated in the tool-tip.

Technical notes:
If you set the "enabled" param of a line editor to "false", that will disable text selection in the widget.
However, if a line editor is initially enabled and then gets disabled with setEnabled(FALSE), it will remain selectable.
It looks like bug, but I'm not going to fix it now to not introduce issues like this one.
Instead, I'm applying to workaround which is to enable the widget initially and disable later, thus keeping it read-only but selectable.