STORM-335 FIXED Made group name clickable in About Land > General tab.
Technical notes:
If you set the "enabled" param of a textbox to "false", that will prevent the widget from receiving mouse events.
However, if a text box is initially enabled and then gets disabled with setEnabled(FALSE), it will remain clickable.
This is a long-standing design problem with "enabled" flag being intermixed with "read-only".
I'm working around it by enabling the group name textbox initially, thus making it clickable.