SH-480 FIX meshes are not added to objects folder on upload
curl host scrubbing was causing object creation for mesh prims to fail.
Commenting out the offending lines so we can have a functional release.
Will file a separate bug for Vir to investigate when/if we can re-enable
host scrubbing on these URLs for a future release.

Approach approved by davep.