STORM-721 FIXED Fix for text clipping in a textbox without a scroller.
Committed on behalf of Richard Linden:
This patch makes clipping work consistently with text and embedded widgets. The widgets will only be displayed if the corresponding text they are embedded in is displayed.  There is also a new param "clip" for text widgets that can be used to disable clipping entirely.  I introduced this as a potential work around due to the fact that we *used* to assume that text was never clipped in the vertical direction unless we had scrolling turned on.  This hasn't been the case for over a year, but now we can selectively turn off vertical text clipping with (clip="false") if we have to.