Fix for VOICE-3 [crashhunters] LLVoiceVoiceClient::stateMachine
Race condition with parcel changing while shutting down.  If a parcel
changes, we went into code that requests the cap for the new
parcel voice channel info.  This happened in any state, not
the two states where it is really appropriate.  stateRunning
or stateNoSession.

When shutting down, we may be in stateTerminate, stateLeavingSession,
or so on, and we don't want to muck with getting caps and such during
that time.