SH-3941 SH-3954 FIX local appearance does not update properly
We were using an indicator of whether there *was* a baked texture to whether we
were *using* a baked texture. Switched to one of the (several) proper indicators
of whether a baked texture was in use. This should allow local texture data to
get marked for download when we enter appearance mode.

Local testing confirms that old apperances linger for a second or two, but are
soon replaced by the correct local renders.