SH-4226 FIX wearing petite outfit gives appearance responder errors
Requesting appearance updates on updateGeometry for the avatar was spamming
the back-end and causing the throttling mechanism to get hit. Removing that
caused a re-introduction of SH-4109, so added a callback to link creation
in the COF for attachments to request an appearance update.

This should cause us to request an appearance update once per attachment
attached, where before we were seeing up to 8 redundant requests.