STORM-2017: Added translatable help text to rotation.
I chose the camera’s up vector to place the help text as it provided a consistent location on the screen for the user to see the text pop up.

While doing this I realized that the calls to hud_render_utf8text utilized a condition that was guaranteed to be false based on a surrounding if-statement, and so could trivially be replaced with a constant.
Also cleaned out a compiler warning about unused private member variables in llmaniptranslate.  I don’t like warnings and useless code. :P